api1/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.1
numpy==1.26.0
pandas==2.1.1
spacy==3.7.2
nltk==3.8.1
